+++ README.md
@@ -29,10 +29,9 @@
 
 ### 리팩터링 및 고민 거리
 
-- [ ] 예외 메세지를 하나의 enum 으로 묶기
+- [x] 예외 메세지를 하나의 enum 으로 묶기 (-> exceptionMessage 패키지 내부로 묶음)
 - [ ] 재귀를 계속하다가 StackOverFlowException 이 걸리면?
 - [ ] 도메인과 모델의 차이
-
 - [ ] 사다리 최대 높이 제한 (-> 100 으로 제한)
 - [ ] 플레이어 최대인원을 콘솔 창에 사다리가 깨지지 않고 이쁘게 나올 수 있을 정도로 제한 (-> 26으로 제한)
