+++ src/main/java/ladder/model/Players.java
@@ -1,5 +1,7 @@
 package ladder.model;
 
+import ladder.exceptionMessage.ExceptionMessage;
+
 import java.util.Collections;
 import java.util.List;
 
@@ -16,7 +18,7 @@ public class Players {
 
     private void validatePlayerCount(List<Player> players) {
         if (players.size() < MIN_PLAYER_COUNT) {
-            throw new IllegalArgumentException(ErrorMessage.EXCEPTION_INVALID_PLAYER_COUNT.getMessage());
+            throw new IllegalArgumentException(ExceptionMessage.EXCEPTION_INVALID_PLAYER_COUNT.getMessage());
         }
     }
 
@@ -28,16 +30,4 @@ public class Players {
         return Collections.unmodifiableList(players);
     }
 
-    private enum ErrorMessage {
-        EXCEPTION_INVALID_PLAYER_COUNT("게임을 진행하기 위해서는 두 명 이상의 플레이어가 필요합니다.");
-        private final String message;
-
-        ErrorMessage(String message) {
-            this.message = message;
-        }
-
-        private String getMessage() {
-            return message;
-        }
-    }
 }
